+++ scripts/lichessTools-init.js
@@ -131,6 +131,7 @@ lichessTools.loadTool(LiChessTools.Tools.ChallengeOptions);
 lichessTools.loadTool(LiChessTools.Tools.BetterBestArrow);
 lichessTools.loadTool(LiChessTools.Tools.DownloadBroadcastPgn);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerMoreGames);
+lichessTools.loadTool(LiChessTools.Tools.CustomBoardImage);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);